New handling of tool time-out error.
Now, when AppVet times-out waiting for a report from one or more tools,
it will disable the tool adapter for those tool(s) still in the
SUBMITTED state and display an alert to the user. In addition, an email
is sent regarding this issue to Administrators. Note that multiple tools
timing-out may be the result of only one tool that is taking up too many
resources on the tool server. To avoid one tool affecting the results of
other tools, it is critical to place each tool on a separate server.